+++ Commands/lib/aphid/rake.rb
@@ -1,0 +1,5 @@
+# encoding: UTF-8
+
+require File.join(File.dirname(__FILE__), "extensions")
+require File.join(File.dirname(__FILE__), "rake", "support")
+require File.join(File.dirname(__FILE__), "rake", "tasks")